Added Config Parameter to control STS check on startup (#70)
*Issue #, if available:*

Closes https://github.com/aws/aws-secretsmanager-agent/issues/36

*Description of changes:*

This pull request adds a new configuration option called sts_validation
. This option lets users turn off the STS GetCallerIdentity check that
happens when the AWS Secrets Manager agent starts up. By default, the
check is enabled to ensure credentials are valid when the agent
initializes and to maintain compatibility with existing setups. However,
this new parameter gives users the flexibility to skip the validation in
situations where user does not want it.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Simon Marty <martysi@amazon.com>