Pin to rustls version 0 (#93)
*Issue #, if available:*

*Description of changes:*

Crates.io requires all crate dependencies to be pinned to a version
number. `*` dependency versions are not allowed. This causes the
publishing the latest update of the crate to fail since it has a
dependency on `rustls = "*"`.

```
Caused by:
  the remote server responded with an error (status 400 Bad Request): wildcard (`*`) dependency constraints are not allowed on crates.io. Crate with this problem: `rustls` See https://doc.rust-lang.org/cargo/faq.html#can-libraries-use--as-a-version-for-their-dependencies for more information
```

Bumped version number to 1.2.1

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.